# Cap the amount of buffering done by BytesPipe