Previously, BytesPipe would grow without a limit. This was dangerous if
data was written to it faster than it was read, it would exhaust available
memory.

Turn BytesPipe's Read and Write functions into blocking, goroutine-safe
functions. Add a CloseWithError function to propagate an error code to
the Read function.

Adjust tests to work with the blocking Read and Write functions.

Remove BufReader, since now its users can use BytesPipe directly.

This version of the commit hardcodes a 1 MB threshold for blocking on writes. I'm hoping that we can do some testing around this and determine if this number should be tweaked or made configurable.

cc @unclejack @vcaputo @jonathanperret