@unclejack 2015-11-12
@aaronlehmann There seems to be some kind of problem. The benchmark doesn't seem to finish right now. `BenchmarkBytesPipeRead-2` hangs forever for no apparent reason for me. This is something I can reproduce every single time.

Can you look into this, please?