@LK4D4 2015-11-12
I wonder how that behchmark worked before :) It writes 500 1024 bytes chunk, but reading 1000 and it has check that n is 1024 each time.