@aaronlehmann 2015-11-12
Sorry about that - I missed the benchmarks.

As @LK4D4 mentioned, the benchmark is trying to read more data than is written. With the new blocking behavior, this causes a deadlock (not that it would have worked before, either).

I've fixed this by making the benchmark read one thousand 512 byte chunks instead.