@aaronlehmann 2015-11-12
BTW, here is the change in benchmark results:

before:
```
BenchmarkBytesPipeWrite	   10000	    172635 ns/op
BenchmarkBytesPipeRead	   30000	     40503 ns/op
```

after:
```
BenchmarkBytesPipeWrite	   10000	    197825 ns/op
BenchmarkBytesPipeRead	   10000	    127205 ns/op
```

I think the worse nubers in the "read" benchmark reflect the locking that was added. However, this doesn't reflect a real-world slowdown, because formerly `BytesPipe` had to be used through `BufReader`, which implemented locking on top of it. I suspect performance of the new `BytesPipe` implementation at least matches the old `BytesPipe` + `BufReader`. It might even be significantly better, since it avoids an unnecessary goroutine running `io.Copy`.