@unclejack 2015-11-13
The fact that it's slightly slower doesn't even matter. The whole point is to avoid the memory issues and have good performance while doing that.

LGTM